fixed get_anrw_options + moved debug functions to history_debug
