+++ includes/history/history_debug.h
@@ -1,0 +1,9 @@
+#include <libft.h>
+#include "shell_env.h"
+#include "history_options.h"
+#include "range.h"
+
+// #define HISTORY_DEBUG
+
+void			print_history_options(t_hist_opt *options);
+void			print_range(t_range *range);
