+++ set_histfile.sh
@@ -1,2 +1,3 @@
 echo "one\ntwo\nthree\nfour\nfive\nun deux trois quatre cinq" > .42sh_history
 echo "one\ntwo\nthree\nfour\nfive\nun deux trois quatre cinq" > tests/.42sh_history
+rm tests/test
